[landing] Reduce footer width for "legal" pages
Summary: We currently reduce the header width when we're on one-column "legal pages" to match the width of the text. This diff matches that for the footer.

Test Plan:
Here's what it looks like: https://blob.sh/atul/5a80.png
(Note: the sitemap colors and "request access" button responsiveness are "known issues" and on the to-do list)

Reviewers: ashoat, palys-swm

Reviewed By: ashoat, palys-swm

Subscribers: palys-swm, Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D2140