[server] Modify `landingHandler:AssetInfo` to include `fontURLs` instead of `fontsURL`
Summary: The duospace font we want to use is not available on Google Fonts. As a result, we'll need multiple font URLs. This diff refactors `landingHandler` in place to handle an array of font urls instead of a single one.

Test Plan: Continues to work as expected & checked "computed" fonts in dev tools.

Reviewers: ashoat, palys-swm

Reviewed By: ashoat, palys-swm

Subscribers: palys-swm, Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D2142