[native] Disable interactions with dummy chat input bar
Summary: We were able to focus the hiddent input bar and open a keyboard while a tooltip was open. The solution is to disable pointer events of the dummy input bar.

Test Plan: Click next to the message while the tooltip is open and verify that the keyboard no longer appears. Also, check if rendering a view enclosing a regular input bar doesn't affect its layout.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: Adrian, atul, karol-bisztyga, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2249