[CommCoreModule] Pull deletion of media out of `SQLiteQueryExecuter:RemoveMessages*`
Summary:
(addresses https://phabricator.ashoat.com/D2256?id=6487#inline-10875)

Previously, we were handling the deletion of media with the deletion of messages in `SQLiteQueryExecutor`

However, we concluded that the functions in `SQLiteQueryExecutor` should modify the DB in precisely the way they're named and not delete things on other tables as a "side effect".

So, we introduced `removeAllMedia`, `removeMediaForMessages`, and `removeMediaForThreads` in `SQLiteQueryExecutor` and called them along with the corresponding message removal functions in `execute()` of each operation.

Test Plan:
Haven't yet but
1. Will insert messages and media into client DB
2. Remove messages via `remove` message store op
3. Make sure the media associated with the removed messages are also deleted.

Reviewers: ashoat, palys-swm

Reviewed By: ashoat

Subscribers: palys-swm, Adrian, karol-bisztyga, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2261