[services] Exctract base image
Summary: In D2207 Tomek pointed out that the backup base image is built on the top of the tunnelbroker base image. As those services are separate, I extracted a base image for all the future services. This image installs grpc(and all necessary packages) as grpc is a base for communication across all the services.

Test Plan: perform scripts from `services/package.json`

Reviewers: palys-swm, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, palys-swm, Adrian, atul, benschac, boristopalov

Tags: #aws_init

Differential Revision: https://phabricator.ashoat.com/D2243