[CommCoreModule] Introduce `processThreadStoreOperationsSync` core function
Summary:
Introduce synchronous version of `processThreadStoreOperations`, that returns boolean instead of promise. It will be used in redux-persist store migration.
C++ thread is blocked until database operations are finished.

Test Plan: The function is not used in code yet, but I tested it and it works fine on virtual iOS device.

Reviewers: palys-swm, atul, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D2196