[native][fix] Fix the 'prepareGrpc' task up-to-date state due to the zconf.h modification.
Summary:
Fix the 'prepareGrpc' Gradle task not UP-TO-DATE state due to the 'zconf.h' modification in the 'third-party' gRPC modules.
More about the problem at the [[ https://www.notion.so/commapp/Gradle-task-app-prepareGrpc-does-some-work-every-compilation-660051f154994b7682c1c6008b77ee46 | Notion task ]]

As this problem can occur later in some other new packages I've added the new parameter `runAfter` to the Gradle custom cloning task.
This parameter is a list of the shell commands which needs to be run after the cloning process.
 Adding the remove of the `/third_party/zlib/zconf.h` command to the runAfter solved the problem.

Test Plan:
Tested using `./gradlew installDebug` and `yarn react-native run-android` commands.

Expected results:
Task is in UP-TO-DATE state and don't executes on the second run.

Results:
Task is not executed and in UP-TO-DATE state on the second run and beyond if nothing is changed on the inputs.

Reviewers: ashoat, palys-swm, karol-bisztyga

Reviewed By: ashoat

Subscribers: palys-swm, Adrian, atul, karol-bisztyga, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2273