[CommCoreModule] Introduce NetworkModule
Summary:
Create a new module that will handle all the networking logic. This module will be called from a global singleton (the next diff).
This whole stack is based on a discussion from a couple of diffs, from which D2285 is the most significant.

Test Plan: After this diff - no changes. Tested with the next diffs.

Reviewers: karol-bisztyga, ashoat

Reviewed By: karol-bisztyga, ashoat

Subscribers: ashoat, Adrian, atul, benschac, boristopalov

Tags: #background-notifications

Differential Revision: https://phabricator.ashoat.com/D2312