[CommCoreModule] Send pong when background notif is received
Summary: When a backgroung notification is received, we need to call a pong method. In the future we would have more background notification types, so we will choose the right action based on the type.

Test Plan: Modify background code to send background notifs and check if the text is logged to the console.

Reviewers: ashoat, karol-bisztyga

Reviewed By: ashoat

Subscribers: Adrian, atul, benschac, boristopalov

Tags: #background-notifications

Differential Revision: https://phabricator.ashoat.com/D2315