[lib] Fix `existingThreadInfoFinder` to find threads that changed type after being pending
Summary:
Solution for https://linear.app/comm/issue/COM-86/after-sidebar-to-subthread-promotion-user-can-navigate-back-to-non and https://linear.app/comm/issue/COM-85/not-delivered-message-in-pending-sidebar-is-lost-when-sidebar-upgraded bugs.
The problem existed, because `existingThreadInfoFinder` couldn't match newly created subthread, that was initially sidebar to old pending sidebar name (the one "/pending/type5/...") which was still in app navigation route.
To fix it, I made two changes:
- modified `locallyUniqueToRealizedThreadIDsSelector` to include all possible "pending" names, basing on introduced function `getPossibleInitialThreadTypes` that returns list of thread types, which thread could be in the past.
- In `useExistingThreadInfoFinder` replaced checking condition if thread type can be pending to cheking if it could be in the past (like with the subthread, that initially could be pending sidebar).
	This condition checks if current thread type can be pending, or thread contains source message, so it does not affect subthreads that are created directly (not by upgrading sidebar).

Test Plan: I tested the solution on both private and public subthreads with cases described in D2294 and it solves both encountered problems.

Reviewers: palys-swm, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2319