[CI] Automated android releases via GitHub Actions
Summary:
To make a release after this diff:
1. Tag `codeVersion` commit as usual
2. Select "Make a release" from GitHub web interface and select the tag from the dropdown.
3. This "action" should trigger on the release being "published" (eg hitting the green button)

Note: I've jotted down notes along the way. I'll clean them up and put them on Notion once things are fully figured out.

Test Plan: Build 109 was fully built for release and uploaded using this action on a fork of this repo.

Reviewers: ashoat, palys-swm

Reviewed By: ashoat

Subscribers: palys-swm, Adrian, karol-bisztyga, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2324