[docs] Removed brew installation of Node, included Node installation in nvm configuration
Summary:
Looked into the issue raised in Notion macOS/Xcode updates, in which there was an issue with `sqlite3` during `yarn cleaninstall`, caused by Node.js default being version 17.0.

I removed the section in which Node is installed using Homebrew, and included Node installation steps in the nvm configuration.
- I moved the section explaining why we use `brew install <program>; brew upgrade <program>` into the PHP config section because I got rid of the Node config section. I think this explanation is still worthwhile for context
- Additionally, I put a line showing the command used to change the default Node used by nvm, so users don't have to keep running `nvm use <Version>` before `yarn dev`.

This change should streamline the Node installation process, and provides a command with which users can change their default Node.js version when we update it in the future.

Test Plan: N/A

Reviewers: atul, ashoat

Reviewed By: atul, ashoat

Subscribers: ashoat, palys-swm, Adrian, karol-bisztyga, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2360