[native] Don't include `loop` in mediaUpdate payload for `Image`s
Summary:
The `updateMultimediaMessageMediaActionType` action being dispatched from input-state-container:709 includes a `loop` property in the mediaUpdate payload for both Image/Video media.

In message-reducer:1247, this `mediaUpdate` is merged with `singleMedia` for media of type `Image`.

However, the `Image` type does not include a `loop` field, which means the object in redux is no longer a valid `Image`. This leads to flow issues with the `translateMediaToClientDBMediaInfo` helper function in `message-ops-utils` which expects an input of type Media (Image | Video).

Excluding the `loop` property in the mediaUpdate payload for media of type `Image` resolves the issue.

Test Plan:
Before:
```
Cannot call `translateRawMessageInfoToClientDBMessageInfo` with `rawMessageInfo` bound to `rawMessageInfo` because property `loop` is missing in `Image` [1] but exists in object literal [2] in array element of property `media`. [prop-missing] (References: [1] [2])
```

After: no more flow errors

Reviewers: ashoat, palys-swm

Reviewed By: ashoat

Subscribers: palys-swm, Adrian, karol-bisztyga, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2375