[lib] Add `messageContentForClientDB` to `MessageSpec` type
Summary:
For each message type, we'll implement a `messageContentForClientDB`  function to give us the content of the `content` column in the SQLite `messages` table.

For most of the message types, this function will pretty much be an alias for `messageContentForServerDB`, but there are some that are a bit more complicated (eg `unsupported` and maybe `sidebar_source`).

Test Plan: NA, will test the implementations though.

Reviewers: ashoat, palys-swm

Reviewed By: ashoat

Subscribers: palys-swm, Adrian, karol-bisztyga, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2406