[lib] Implement `messageContentForClientDB` where alias of existing `messageContentForServerDB`
Summary:
This diff basically implements `messageContentForClientDB` as alias of `messageContentForServerDB` for message types that have existing implementations of the latter.

So what's the point?
There are message types where we want to store something different in the `content` column of the SQLite DB than on the MySQL SB.

One example is the `unsupported` message type which we need to persist in the SQLite store, but which we don't store at all in the MySQL DB.

Test Plan:
1. Fill SQLite DB with some messages
2. Make sure the `content` columns look as expected

Reviewers: ashoat, palys-swm

Reviewed By: ashoat

Subscribers: palys-swm, Adrian, karol-bisztyga, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2410