[2/n] [services] Add MultiPartUploader
Summary:
There was a problem with appending to the S3 object. S3 doesn't handle mapping in place so we'd have to first download the entire object, then store it in some variable, append to it and write again to the cloud. That solution doesn't scale.

What gave me hope about this was the concept of "multipart upload" which is possible using AWS S3.

Here is the code for a handler for multi-part uploads, more context is available here: https://linear.app/comm/issue/ENG-171/figure-out-appending-to-s3-objects

Test Plan:
For now, I tested this with a file prepared specifically for that. I'm going to spend some time figuring out how that can be tested. The main problem here is that we need to simulate a situation when there's over 5MB of data in a single file - it is totally doable, just takes a long time to, for example, generate that amount of random bytes.

So, I'm leaving this test plan as TODO but wanted to upload the code so it can be reviewed.

Reviewers: palys-swm, geekbrother, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, palys-swm, Adrian, atul, benschac, boristopalov

Tags: #aws_implement_services

Differential Revision: https://phabricator.ashoat.com/D2404