[native] Test iOS notif badging theory for staff-only build
Summary:
I have a theory that the notif badging issue is happening as a result of `PushHandler` not being rendered because `await SplashScreen.hideAsync()` never resolves.

This diff is meant for a staff-only build to test my theory. We should revert this diff after putting the build out, so we avoid having this `Alert` code on `master`.

If my theory is correct, I expect two differences in behavior:

1. We should sometimes see the `Alert` in question appear
2. After 5 seconds of waiting, the `PushHandler` should get rendered regardless of whether `await SplashScreen.hideAsync()` resolves. So after seeing the `Alert`, the badging issue should disappear

Test Plan: Test in prod!

Reviewers: atul

Reviewed By: atul

Subscribers: palys-swm, Adrian, karol-bisztyga, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2421