[native] Update badge count when push handler is mounted
Summary: The issue and investigation is described in https://linear.app/comm/issue/ENG-99/badge-count-issue-on-ios. The root cause is that the badge count should be updated when push handler is not yet mounted. The solution is to update the count when the handler mounts.

Test Plan: Introduce a sleep in app-navigator so that push handler is not rendered for some time during app start. Send a message from another client which should result in badge count increase. Notice that the badge count stays the same during the sleep and the is increased when the timeout elapses.

Reviewers: ashoat, atul

Reviewed By: ashoat

Subscribers: Adrian, karol-bisztyga, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2423