[native] Update execute call in settings.gradle
Summary: I was seeing an error "index 0, length 0" printed by the method call. Checked the latest install docs for `react-native-unimodules` (which is what requires these lines) and noticed that the execute calls had params: `execute(null, rootDir)`. Made that change and it solved the issue.

Test Plan: This make it possible to build the Android app on my new M1

Reviewers: atul, anand

Reviewed By: anand

Subscribers: palys-swm, Adrian, karol-bisztyga, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2430