[server] Reduce delete updates batch size
Summary: When creating a new thread we perform work proportional to number of users in parent thread squared, which makes so it times out. We should definitely address it, but it made another issue visible. The current batch size was so big it was impossible for the query to finish before the timeout elapsed. Reducing the batch size to 500 reduces the time needed for creating a thread with parent with 80 members from 20s to 2s on my local. Although this isn't the solution of the main issue, we shouldn't have the batch size that is guaranteed to fail.

Test Plan: Create a new thread and measure the time. Reduce barch size and create another task. Compare the times.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: Adrian, atul, karol-bisztyga, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2475