[server] Make multimedia message creation idempotent
Summary:
Linear issue: https://linear.app/comm/issue/ENG-236/failure-when-sending-image

`createMessages` itself is idempotent, but `fetchMedia` will throw if the media in question has already been assigned a container.

This solution is a bit hacky because we end up calling `fetchMessageInfoForLocalID` twice (also called inside `createMessages`), but I don't have time to refactor it to be better and I think we should patch this bug.

Test Plan: My dev environment is almost set up but not quite... so I'm sort of relying on careful reading and Flow right now

Reviewers: atul

Reviewed By: atul

Subscribers: palys-swm, Adrian, karol-bisztyga, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2476