[1/2][CommCoreModule] Add option to close network module
Summary: The best way to close a grpc connection is to call a destructor of the client and to do that we just need to reset the pointer to our network Client.

Test Plan: It's hard to test it well now as we're not initializing the connection yet, so the proper testing will be performed later. For now checking if the app can be run should be enough.

Reviewers: karol-bisztyga, ashoat

Reviewed By: karol-bisztyga, ashoat

Subscribers: ashoat, Adrian, atul, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2540