[2/2][CommCoreModule] Close old connection when enabling multithreading
Summary:
When we switch from single to multithread we would like to close the old grpc connection.
All the interactions with network module goes through network singleton which means that we can safely make the old network module unusable as all the new calls will be handled by a new thread_local instance.

Test Plan: Add a console log to close method and check if it is called when we start the app.

Reviewers: karol-bisztyga, ashoat

Reviewed By: karol-bisztyga, ashoat

Subscribers: ashoat, Adrian, atul, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2541