[server] Group conditions when deleting updates
Summary:
When we create a new thread, a query with a lot of delete conditions is executed. The conditions usually differ only slightly so we can replace `OR` with `IN`, e.g. replace `(C1 AND C2 AND C3 AND X1) OR (C1 AND C2 AND C3 AND X2)` with `(C1 AND C2 AND C3 AND X IN (X1, X2))`. By doing that we improve the performance, because IN is slightly faster than OR when dealing with long conditions, but more importantly, we reduce the size of the query and number of variables sent to the db.

There's more place for further improvements. If the performance is still unacceptable, we can consider introducing more `IN` statements, but figuring out dynamically the optimal set of conditions is a lot harder task.

Test Plan: Modify the code to produce both old and new query and check if they represent the same condition by checking if there's one to one relationship between them.

Reviewers: atul, def-au1t, ashoat

Reviewed By: atul, ashoat

Subscribers: ashoat, Adrian, karol-bisztyga, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2550