[docs] Fixing small nit in "Android emulator" section. Android 10 to Android 11.
Summary: "Android emulator" section was out of date. New installations of latest Android Studio seem to come with an emulator pre-installed, but I think it's a good idea for devs to download an emulator to implicitly double-check for Play store compatibility. Fixed a double "the" that I saw. Changed instructions to suggest Android 11 for system image version.

Test Plan: Downloaded a Pixel 3 with Play Store support, downloaded app via Play Store. Separately, I built a debug build and ran it on the emulator.

Reviewers: atul, ashoat

Reviewed By: atul, ashoat

Subscribers: ashoat, palys-swm, Adrian, karol-bisztyga, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2560