[landing] fix header request access button
Summary:
Fixing the header on the landing page. This is a bit of a compromise. The header looks busy because...it is. We don't have a true tablet header so we can really fit the full button with padding. I think it's a weird experience to see a request access button head re-size dramatically.

https://linear.app/comm/issue/ENG-316/header-button-needs-to-be-contained-to-one-column-per-design

{F6042}

Test Plan: atul

Reviewers: atul, ashoat

Reviewed By: atul, ashoat

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2635