[server] Extend cookie lifespan only when a request was not a ping
Summary:
Every socket message is causing an update in the db. Every connected client sends a ping every 3s which resulted in socket lifespan update to be one of the most often blocked queries https://linear.app/comm/issue/ENG-308/check-server-logs-for-query-improvements. We don't need to update this with such a precision as cookie lifetime is 30 days.
The first solution I considered is to throttle the update function, but we're running our server concurrently and it could be hard to do that reliably.
A better solution is to just ignore PING message when updating the lifetime. Every meaningful action will still extend the lifetime, but if someone kept an app opened for 30 days without any interaction we would mark the cookie as outdated... but I don't think it is ever possible to happen.

Test Plan: Modify `extendCookieLifespan` and check if PING does not extend the cookie but every other interaction does it.

Reviewers: atul, def-au1t, ashoat

Reviewed By: ashoat

Subscribers: ashoat, Adrian, karol-bisztyga, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2652