[4/n] [services] Backup Dev Mode - add code
Summary:
In response to https://linear.app/comm/issue/ENG-309/create-dev-mode-for-docker-backup-server-service

This is built on top of D2617 D2618 D2619

Adding code for the dev mode for the backup service.
In the dev mode, we do not use AWS S3, we use the local file system.

Test Plan: Please see D2630

Reviewers: geekbrother, anand, palys-swm, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, palys-swm, Adrian, atul, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2629