[5/n] [services] Backup Dev Mode - add configurations
Summary:
In response to https://linear.app/comm/issue/ENG-309/create-dev-mode-for-docker-backup-server-service

This is built on top of D2617 D2618 D2619

Adding configurations for the dev mode of the backup service.

For now, I only implemented the dev mode in the backup service. Tunnelbroker does not really need it - the dev mode would not differ from the production mode, but adding the dev mode to any other service will not be a problem. All we will need is to configure CMake a little bit and add custom dev implementation of the desired functionalities.

Test Plan:
```
cd services
run-backup-server-dev-mode
test-backup-server-dev-mode
test-all-services-dev-mode
run-all-services-dev-mode
```

Reviewers: palys-swm, geekbrother, anand, ashoat

Reviewed By: palys-swm, anand, ashoat

Subscribers: ashoat, palys-swm, Adrian, atul, benschac, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2630