[CI] Skip iOS/Android build for changes in irrelevant directories
Summary:
Realize this isn't a priority but this was a super quick change.

Just avoiding iOS/Android builds for changes to `landing`, `web`, `docs`, and `server` since I don't think changes to those directories could change the outcome of native builds.

One day in the future `server` might be relevant if we're doing integration tests or whatever, but I think we can exclude it for now.

Relevant section in GH Actions docs: https://docs.github.com/en/actions/learn-github-actions/workflow-syntax-for-github-actions#example-ignoring-paths

Test Plan: Tested on a fork of repo

Reviewers: palys-swm, def-au1t, varun, ashoat

Reviewed By: varun, ashoat

Subscribers: ashoat, Adrian, karol-bisztyga, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2705