[3/n] [id-schemas] [server, lib] Move part of `validation-utils` into `lib`
Summary: Move `validation-utils` functions, that do not depend on server, to `lib` folder, so the validation utils could be used in lib. The goal is to put type validators next to the types in `lib/types` in next diffs.

Test Plan: Project builds and runs as before. App works correctly on virtual iOS device.

Reviewers: palys-swm, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: benschac, ashoat, Adrian, atul, karol-bisztyga, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2730