[CI] Only build `arm64-v8a` for Android pre-land Build CI
Summary:
Only build for `arm64-v8a` when doing an Android CI build "pre-land" to speed things up. We use the `BUILDKITE` env variable to determine that this is a pre-land build. The `CI` environment variable is set to true both on buildkite and GH Actions so we can't use that. Sure, if we were to use BuildKite for "post-land" CI builds we'd need to change this, but I think it's fine for now.

We will still build all architectures after a commit has landed via GH Actions which should catch any architecture-specific issues that arise... which I think is unlikely

Test Plan: See how much time is saved when the "pre-land" CI kicks off

Reviewers: geekbrother, palys-swm, varun, ashoat

Reviewed By: ashoat

Subscribers: benschac

Differential Revision: https://phabricator.ashoat.com/D2743