[6/n] [id-schemas] [lib] Introduce `RawMessageInfo` validator
Summary:
Introduce validators for `RawMessageInfo` and all it's subtypes. As there are 19 types of messages and they are very similar, I didn't want to create separate 19 diffs for the changes.
However, I can create separate diffs for each type if there is a need.
Test Plan:
I tested the validators with `jest` tests, that I will add in the next diff.
Reviewers:
palys-swm
Subscribers: