[native] Throw an exception when closed network module is initialized again
Summary: We're closing the network module only when switching from single to multi thread and we should have only one network module active at a time. The fact that closed module is initialized again means that we're initializing the instance from the main thread, which we should never do.

Test Plan: Modify Global Network Singleton code to call initialize after network module is closed and verify that an exception is thrown

Reviewers: karol-bisztyga, atul