[native] Naming convention for "MultimediaMessageTooltipHeight" vs "multimediaMessageTooltipHeight"
Summary: While looking into ENG-160, found this naming inconsistency, where `multimediaMessageTooltipHeight" was not Pascal cased (?). I noticed that the other export `MultimediaMessageTooltipModal` was capitalized, and naming convention between `multimedia-message` files otherwise appear to conform to a single convention.

Test Plan: This renaming is contained to the two files listed, and does not introduce new program behavior. I tested the tooltip in simulator to ensure nothing broke with this change.

Reviewers: atul

Subscribers: