[web] [15/n] add some base css utils
Summary:
This might be better to have as a conversation if there are strong opinions. I started grabbing functional utility classes that we need from https://github.com/basscss/basscss. The reason I didn't grab the library from NPM is:

- there isn't any documentation for grabbing the classes we need. (flex, text-align, etc.)
- there's a bunch of CSS we don't need and I didn't want to start configuring webpack with CSS purge.
- It's a really easy way for a developer at a glance to know which utility classes we have so they don't need to duplicate code.
- base's final release was a couple of years ago and there aren't plans to change it. They're just CSS classes.

(Note, I didn't add everything we need just yet because this is a rather large change to how we've been doing things and I want to know you're opinions)

base was the pre-cursor to tailwind and other CSS utility framework classes.

So, why not tailwind?
- I don't want to spend the time configuring it when most of what we need is in base.
- We have our own style guide that doesn't look like tailwind or match its spec.
- It's a bit more of a learning curve IMO.

So, why do it this way at all?

- I wanted to show y'all what I was thinking and how I've written larger CSS projects that don't use CSS-in-JS. We did this at https://rundexter.com about 5 years ago before CSS-in-JS took off. Their design style guide used basecss's defaults for the most part and overwrote what didn't fit. In general, I think this is a valid approach but isn't great. It makes understanding our CSS at glance possible in just the component and is highly composable. We end up getting a consistent system where 90% of repeated styles can be expressed in functional classes and mostly bespoke CSS modules are needed.

Alternative:
- Is to use a CSS-in-JS solution like styled or emotion. For the record, I'd use something like https://vanilla-extract.style, but flow. If you want to see this design system in action check out mirror: https://github.com/mirror-xyz/degen
- We'd have strict typing and auto complete.
- We'd just import components like `<Layout.Row align="center">Hello</Layout.Row>`, rather than repeating classes repetitively.
- It's super easy to find dead code we don't use with eslint, prune and flow.
- It comes at a perf cost. CSS-in-JS libs ship with a run time.
- We could possibly share components with native. But, also perf. I can't find it but someone at rainbow had a good tweet about using styled. I'm assuming they're using vanilla extract since the author of the library works on design systems at rainbow.
- Theming is super easy.

Just for the record I think both approaches are valid, I don't have a hard bias for either. Additionally, if anyone else has a different approach they'd want to take. From what I see it looks like the community has embraced CSS-in-JS. Tailwind is pretty incredible esp. with autocomplete support in vscode. Base doesn't have auto-complete but is still a great mental model to work with.

Test Plan: N/A

Reviewers: atul, ashoat

Reviewed By: atul, ashoat

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga, boristopalov

Differential Revision: https://phabricator.ashoat.com/D2805