[native] Restrict ComposeThread to subchannels
Summary:
See my comments about `ComposeThread` in [this Linear comment](https://linear.app/comm/issue/ENG-300#comment-a004e32a).

The changes in the previous diff introduced an issue in cases where `ComposeThread` is used without a parent thread. In fact, we don't currently use it that way anywhere in the app, since we moved parentless thread creation to a new flow. So I figured we could make that restriction official, to avoid the potential erroneous usage and to simplify the existing code.

Test Plan: I haven't had a chance to test this

Reviewers: marcinwasowicz, palys-swm

Reviewed By: marcinwasowicz, palys-swm

Subscribers: benschac, Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D2840