[server] Send a background notif to rescind notifications on native
Summary: Use a background notif that contains an id of a notification that should be removed

Test Plan:
Logged in as user A on a device and on the web and as a user B on the web. As B, sent a message to a thread which resulted in A receiving a notification on the device (the app was backgrounded). As A on the web marked that thread as read which resulted in A on the device to receive a background notification which removed the notification from the center. Also added some logs to verify that the background notif was received and we were able to find the notification in the center.

I also tested it when the app was closed.

Additionally, the badge count in both cases was set correctly.

Reviewers: karol-bisztyga, marcinwasowicz, ashoat

Reviewed By: karol-bisztyga, ashoat

Subscribers: benschac, ashoat, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D2892