[web] [refactor] move user settings css to its own file
Summary: Only moving this code because I need to edit the styles a lot and isolating it to its own file will speed things up {F10088}

Test Plan: make sure css matches what's in the branch before. It still looks a bit off from when I moved logout to the modal. I don't think it's worth nitting at right now because we're going to be changing around the styling

Reviewers: atul, ashoat

Reviewed By: atul, ashoat

Subscribers: ashoat, palys-swm, Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D2937