[web] update chat thread design to more closely match the design of the figma.
Summary: This isn't 100%, but I think it's close enough without having feature parity / a lot of css that needs to be cleaned up. <add screenshot> <add screenshot>

Test Plan: generally looking for feedback if we can call this done for the moment while the rest of the thread features are built.

Reviewers: atul

Subscribers: ashoat, palys-swm, Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D2944