[web] update chat thread design to more closely match the design of the figma.
Summary:
This isn't 100%, but I think it's close enough without having feature parity / a lot of css that needs to be cleaned up.
in diff: {F10154}

actual: {F10158}

https://www.figma.com/proto/a1nkbWgbgjRlrOY9LVurTz/Comm-%2F-Desktop-app?page-id=1170%3A77104&node-id=1170%3A79329&viewport=459%2C48%2C1.87&scaling=min-zoom&starting-point-node-id=1170%3A77116

Missing: bread crumbs, avatar / pfp upload, arrows and community inset (ex. swm in top right corner).

Test Plan: generally looking for feedback if we can call this done for the moment while the rest of the thread features are built.

Reviewers: atul, ashoat

Reviewed By: atul, ashoat

Subscribers: ashoat, palys-swm, Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D2944