[RN0.66] 4/n: Introduce Android rn_edit_text_material.xml Drawable
Summary:
This change is actually for React Native 0.67. I started this upgrade with the goal of updating to 0.67, but was stymied. Instead I updated to 0.66, but I figured I would leave this change in so we don't have to do it once we update to React Native 0.67.

Context [here](https://github.com/facebook/react-native/commit/254493e1fb0c3a1e279e2c957e83edac6252d041).

Note that I omitted the change to `native/android/app/src/main/res/values/styles.xml`, and as a result this Drawable is not being used. I'm open to excluding it, but I'm hoping that unused `Drawable`s won't be bundled, and figure it's better to keep ourselves compatible with the React Native template to make changes easier in the future.

The reason we are omitting the change to `native/android/app/src/main/res/values/styles.xml` is because we've already overwritten `android:editTextBackground` to avoid quirky Android behavior, and as a result the bugfix that this addresses is not relevant to us.

Test Plan: N/A, see above

Reviewers: palys-swm

Subscribers: Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D2983