[RN0.66] 11/n: Update to Reanimated 2.3
Summary:
It's necessary to update Reanimated versions as of React Native 0.66; see [here](https://github.com/software-mansion/react-native-reanimated/issues/2456).

Note that the next diff is also necessary to get the new version of Reanimated to work with the new version of React Native.

Test Plan: Built and ran the whole stack on iOS and Android across various architectures

Reviewers: karol-bisztyga, palys-swm, atul

Subscribers: Adrian

Differential Revision: https://phabricator.ashoat.com/D2990