[RN0.66] 25/n: Change document.hasFocus check to satisfy Flow
Summary:
Flow seems to want it this way now. It complains whenever it sees `document.hasFocus` without a `()` directly following.

Technically this isn't the same, but it's the same for our purposes. It would technically break if `document.hasFocus` were set, but not set to a function.

Test Plan: Flow

Reviewers: benschac, palys-swm, atul

Subscribers: Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D3004