[CommCoreModule] Hold reference to `jsInvoker` in `GRPCStreamHostObj`
Summary:
We add a `jsInvoker` reference member variable to `GRPCStreamHostObj` and set it in the constructor.

Having a reference to `jsInvoker` lets us queue things up to run on the JS thread. This relevant for the `OnReadDoneCallback` lambdas we will construct in `GRPCStreamHostObj` and pass to `NetworkModule` -> `Client` -> `ClientGetReadReactor`

Test Plan: Tested implicitly in subsequent diffs.

Reviewers: palys-swm, def-au1t, ashoat, karol-bisztyga

Reviewed By: ashoat

Subscribers: Adrian, benschac

Differential Revision: https://phabricator.ashoat.com/D3042