[web] [refactor] add button variant style to account for other btn styles
Summary: We have a couple of different button styles. We need to toggle styles based on the type. first diff of refactor. Going to move it and add another variant style prop + styles. This will be used in modal too

Test Plan: make sure the button left button (gear) is still working correctly

Reviewers: atul, ashoat

Reviewed By: ashoat

Subscribers: ashoat, palys-swm, Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D3039