[server] Set badgeOnly flag when updating only the badge
Summary: We need to set this flag to avoid an empty notification being displayed. We don't have enough info here to do what we do in `prepareAndroidNotification` for older clients, but I guess we can keep it like that. It would be nice if we had some stats about code versions so that we could start increasing a threshold of supported versions.

Test Plan: Mark a thread as read and console log a notif on native.

Reviewers: karol-bisztyga, atul, ashoat

Reviewed By: atul, ashoat

Subscribers: benschac, ashoat, Adrian

Differential Revision: https://phabricator.ashoat.com/D3059