[native] Set badge count on Android in native code
Summary: Move setting the count from js to native. We're using the same functions as `RNFirebaseNotifications` for setting the count. The benefit of this is that we no longer display an empty notification for badge only notifs, as they do not reach the js.

Test Plan: Send a message and mark a thread as read and unread while monitoring the badge count value that is set by the methods. The issue with this is that I was not able to see the badge count on an Android simulator. It looks like this is only supported by vendors https://github.com/leolin310148/ShortcutBadger/issues/174.

Reviewers: karol-bisztyga, atul, ashoat

Reviewed By: ashoat

Subscribers: benschac, ashoat, Adrian

Differential Revision: https://phabricator.ashoat.com/D3069