[docs] Remove g++ from brew install in dev_environment
Summary:
Related linear task: [[ https://linear.app/comm/issue/ENG-697/remove-g-from-brew-install-in-dev-environment | ENG-697 ]]

The `g++` is not provided by the brew anymore. The `brew install autoconf automake libtool curl make g++ unzip` command produces a not found error.
We need to remove the `g++` from the list because it's provided by the code CLI tools which is installed before this command and not present in the brew anymore.

Test Plan: Run `brew install g++`

Reviewers: benschac, jonringer-comm, karol-bisztyga, ashoat

Reviewed By: karol-bisztyga, ashoat

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D3118