[native] Add sendPong method
Summary:
Add a C++ method that can be called from java. This method should be called in a response to PING background notification.
We're still using an old method here, which will be replaced by send soon, but updating it should be trivial.

Depends on D3132

Test Plan: Modify notification handling code to call `NetworkModule.sendPong()`, send a notification and check if the C++ code was called

Reviewers: karol-bisztyga, ashoat, atul, jimpo

Reviewed By: karol-bisztyga, ashoat

Subscribers: jimpo, benschac, ashoat, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D3133