[iOS] Create empty notification service extension
Summary: Add an empty extension using XCode `new target` template. The service extension is executed when the app is backgrounded or killed. When the app is running, this code is not called.

Test Plan:
Modify `prepareIOSNotification` in `send.js` by setting `notification.mutableContent = true`. Modify the code in service so that the `body` gets modified.
Test on a physical device:
1. Run the app and send a message from other client - the notification should be displayed with the original body.
2. Background the app and send a message - the modified notification should be displayed.
3. Kill the app and send a message - the modified notification should be displayed.

Reviewers: karol-bisztyga, atul, jimpo, ashoat

Reviewed By: ashoat

Subscribers: benschac, ashoat, Adrian

Differential Revision: https://phabricator.ashoat.com/D3162