[web] Introduce thread search
Summary:
Introduced thread search feature - the aim was to make it work in exactly the same way as in mobile app.
Instead of using filtered lists from `thread-list-provider` (as it was before), I used the logic very similar to mobile app.
The change also introduces possibility to start new personal thread with other user, as they also appear in search results.
{F9126}
{F13659}

Test Plan: While testing the search, I was comparing it with search results in mobile, and didn't find any differences in any case. Also confirmed, that existing sidebar functionalities works in the same way as before.

Reviewers: palys-swm, atul, benschac, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D2887