Introduce expo splash screen constants to fix 'whie screen' bug on Android
Summary: Following expo migration guide I removed SplashScreen.show(...) call from MainActivity.java. This function is now internally called in expo's code but unless SplashScreenImageResizeMode is explicitely defined in R.string expo uses CONTAIN as default value. This was the cause for the bug which this diff fixes since we used to use NATIVE value. I did not initially put it in R.string since expo migration guide does not advise to do so and it was not necessary when using unimodules.

Test Plan: Build app for android both in debug and releas mode on both emulator and physical device and check that the bug disappears - splash screen smoothly transits to chat list without intermediate white screen.

Reviewers: palys-swm

Subscribers: